requirements.txt
asgiref==3.5.2
certifi==2022.5.18.1
charset-normalizer==2.0.12
Django==4.0.4
django-environ==0.8.1
django-tinymce==3.4.0
flake8==4.0.1
gunicorn==20.1.0
idna==3.3
mccabe==0.6.1
mypy==0.950
mypy-extensions==0.4.3
numpy==1.22.4
psycopg2==2.9.3
pycodestyle==2.8.0
pyflakes==2.4.0
sqlparse==0.4.2
tomli==2.0.1
typing_extensions==4.2.0
urllib3==1.26.9
whitenoise==6.1.0
